refactor error namespacing
